# [release-3.4] fix the potential data loss for clusters with only one member