Backport https://github.com/etcd-io/etcd/pull/14400 to v3.4 fixing https://github.com/etcd-io/etcd/issues/14370

Benchmarking showed 4.5% performance degradation.

Command
```
./bin/etcd --quota-backend-bytes=4300000000
bin/tools/benchmark txn-put --endpoints="http://127.0.0.1:2379" --clients=200 --conns=200 --key-space-size=4000000000 --key-size=128 --val-size=10240  --total=200000 --rate=40000
```

Before 
```
Summary:
  Total:	23.7596 secs.
  Slowest:	0.0911 secs.
  Fastest:	0.0020 secs.
  Average:	0.0236 secs.
  Stddev:	0.0181 secs.
  Requests/sec:	8417.6492

Response time histogram:
  0.0020 [1]	|
  0.0109 [9664]	|∎∎
  0.0198 [130352]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0287 [20841]	|∎∎∎∎∎∎
  0.0376 [5570]	|∎
  0.0465 [1158]	|
  0.0554 [6105]	|∎
  0.0643 [14826]	|∎∎∎∎
  0.0732 [8162]	|∎∎
  0.0821 [2434]	|
  0.0911 [887]	|

Latency distribution:
  10% in 0.0115 secs.
  25% in 0.0129 secs.
  50% in 0.0155 secs.
  75% in 0.0222 secs.
  90% in 0.0593 secs.
  95% in 0.0654 secs.
  99% in 0.0771 secs.
  99.9% in 0.0883 secs.
```

After
```
Summary:
  Total:	24.8914 secs.
  Slowest:	0.1210 secs.
  Fastest:	0.0073 secs.
  Average:	0.0248 secs.
  Stddev:	0.0178 secs.
  Requests/sec:	8034.9071

Response time histogram:
  0.0073 [1]	|
  0.0186 [123685]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  0.0300 [35753]	|∎∎∎∎∎∎∎∎∎∎∎
  0.0414 [5384]	|∎
  0.0528 [5270]	|∎
  0.0642 [19048]	|∎∎∎∎∎∎
  0.0755 [8874]	|∎∎
  0.0869 [1528]	|
  0.0983 [253]	|
  0.1097 [4]	|
  0.1210 [200]	|

Latency distribution:
  10% in 0.0131 secs.
  25% in 0.0141 secs.
  50% in 0.0167 secs.
  75% in 0.0235 secs.
  90% in 0.0592 secs.
  95% in 0.0649 secs.
  99% in 0.0754 secs.
  99.9% in 0.1118 secs.
```